TCB-1 Putting AllowListFilter into its own file
